+++ tests/frames/mod.rs
@@ -8,5 +8,4 @@
  * Documentation: https://nyxspace.com/
  */
 
-mod ephemerides;
-mod frames;
+mod format;
